+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Move "sync axis" and "show only main axis" settings to "Axis" card
 * Add font setttings for measure units and axis
 * Add an ‘auto width’ toggle for category labels that, when enabled, adjusts the label width to display all categories without trimming
+* Add settings to control axis display format and precision
 
 ### Code improvements
 * Fix sync axis by computing min and max category values